backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pytest==8.4.1
httpx==0.28.1
PyJWT==2.9.0
